refactor(stakes): moves storage staking cache models and context props to more generic location to be reused in both notifier and storage
