+++ src/components/organisms/staking/StakingTemplate.tsx
@@ -5,7 +5,7 @@ import { makeStyles, Theme } from '@material-ui/core/styles'
 import { colors } from '@rsksmart/rif-ui'
 import StakingFab from 'components/molecules/staking/StakingFab'
 import React, { FC, useState } from 'react'
-import { StakedBalances } from 'api/rif-marketplace-cache/storage/stakes'
+import { StakedBalances } from 'api/rif-marketplace-cache/common/stakes'
 import { SupportedTokenSymbol } from 'models/Token'
 import DepositModal from './DepositModal'
 import WithdrawModal from './WithdrawModal'
