+++ src/context/Services/staking/index.ts
@@ -1,0 +1,5 @@
+import { Props, State, Action } from './interfaces'
+
+export type StakingContextProps = Props
+export type StakingAction = Action
+export type StakingContextState = State
