+++ src/context/Services/storage/staking/index.ts
@@ -3,24 +3,14 @@ import withStakingContext, {
   Context as StorageStakingContext,
   initialState,
 } from './Context'
-import {
-  Props,
-  Action,
-  State,
-  ContextName,
-} from './interfaces'
-import actions from './actions'
+import { ContextName } from './interfaces'
 
 export {
   StorageStakingContext,
   StorageStakingProvider,
   initialState as storageStakingInitialState,
-  actions as storageStakingActions,
 }
 
-export type StorageStakingContextProps = Props
-export type StorageStakingAction = Action
-export type StorageStakingContextState = State
 export type StorageStakingContextName = ContextName
 
 export default withStakingContext
